MVC e DTO - parte 2

+++ src/main/java/com/projeto/projetojava/services/ProdutoService.java
@@ -41,15 +41,19 @@ public class ProdutoService {
      * @return Retorna um produto caso seja encontrado
      */
     public Optional<ProdutoDTO> obterPorId(Integer id) {
-        //Obtendo optional de produto pelo Id
+
+        // Obtendo optional de produto pelo Id
         Optional<Produto> produto = produtoRepository.findById(id);
-        //Se não encontrar lança Exception
-        if(produto.isEmpty()){
-            throw new ResourceNotFoundException("Produto com id: " + id + " não encontrado")
+
+        // Se não encontrar lança Exception
+        if (produto.isEmpty()) {
+            throw new ResourceNotFoundException("Produto com id: " + id + " não encontrado");
         }
         // Convertendo o meu optional de produto em um produto dto
+
         ProdutoDTO dto = new ModelMapper().map(produto.get(), ProdutoDTO.class);
-        //Criando e retornando um optional de produto dto
+
+        // Criando e retornando um optional de produto dto
         return Optional.of(dto);
     }
 
@@ -60,7 +64,23 @@ public class ProdutoService {
      * @return Retorna o produto adicionado
      */
     public ProdutoDTO adicionar(ProdutoDTO produtoDto) {
-        return produtoRepository.save(produto);
+        // Remover o id para conseguir fazer o cadastro
+        produtoDto.setId(null);
+
+        // Criar um objeto de mapeamento
+        ModelMapper mapper = new ModelMapper();
+
+        // Converter o nosso produto dto em um produto
+        Produto produto = mapper.map(produtoDto, Produto.class);
+
+        // Salvar o produto no banco
+        produto = produtoRepository.save(produto);
+
+        produtoDto.setId(produto.getId());
+
+        // retornar o produto dto atualizado
+
+        return produtoDto;
     }
 
     /**
@@ -69,6 +89,17 @@ public class ProdutoService {
      * @param id do produto a ser deletado
      */
     public void deletar(Integer id) {
+
+        // Verificar se o produto existe
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        // Se não existir lança uma Exception
+        if (produto.isEmpty()) {
+            throw new ResourceNotFoundException(
+                    "Não foi possível deletar o produto com id: " + id + " - Produto não existe");
+        }
+
+        // Deleta o produto pelo id
         produtoRepository.deleteById(id);
     }
 
@@ -79,10 +110,22 @@ public class ProdutoService {
      * @param id      do produto que será atualizado
      * @return Retorna o produto atualizado
      */
-    public ProdutoDTO atualizar(Integer id, ProdutoDTO produto) {
+    public ProdutoDTO atualizar(Integer id, ProdutoDTO produtoDto) {
 
-        produto.setId(id);
-        return produtoRepository.save(produto);
+        // Passar o id para o produto dto
+        produtoDto.setId(id);
+
+        // Criar um objeto de mapeamentp
+        ModelMapper mapper = new ModelMapper();
+
+        // Converter o produto dto em um produto
+        Produto produto = mapper.map(produtoDto, Produto.class);
+
+        // Atualizar o produto no banco
+        produtoRepository.save(produto);
+
+        // Retornar o produto dto atualizado
+        return produtoDto;
     }
 
 }
